Solutions/LeetCode75/TwoPointers/ContainerWithMostWater_test.go
package Solutions

import "testing"

func Test_maxArea(t *testing.T) {
	type args struct {
		height []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			// Input: height = [1,8,6,2,5,4,8,3,7]
			// Output: 49
			name: "test01",
			args: args{
				height: []int{1, 8, 6, 2, 5, 4, 8, 3, 7},
			},
			want: 49,
		},
		{
			// Input: height = [1,1]
			// Output: 1
			name: "test02",
			args: args{
				height: []int{1, 1},
			},
			want: 1,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maxArea(tt.args.height); got != tt.want {
				t.Errorf("maxArea() = %v, want %v", got, tt.want)
			}
		})
	}
}
